fix(quiz): render image answers in pixelate question, tie reveal to timer
