+++ frontend/app/components/quiz/questions/SingleChoice.tsx
@@ -1,15 +1,13 @@
 'use client';
 
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 
 import SubmitButton from '@/app/components/common/SubmitButton';
 import type { QuizChoiceAnswer, QuizImage } from '@/app/types';
-import { cn } from '@/lib/cn';
 
-import ProgressiveQuizImage from '../shared/ProgressiveQuizImage';
 import QuestionCard from '../shared/QuestionCard';
 import QuizLayout from '../shared/QuizLayout';
-import RadioAnswer from '../shared/RadioAnswer';
+import SingleSelectAnswers from '../shared/SingleSelectAnswers';
 
 interface SingleChoiceProps {
   questionNumber: number;
@@ -32,13 +30,6 @@ export default function SingleChoice({
 }: SingleChoiceProps) {
   const [selected, setSelected] = useState<number | null>(null);
   const groupAriaLabel = question.trim() || 'Wybór jednej odpowiedzi';
-  const normalizedAnswers = useMemo(
-    () =>
-      answers.map((answer) =>
-        typeof answer === 'string' ? { text: answer } : answer,
-      ),
-    [answers],
-  );
 
   return (
     <QuizLayout
@@ -52,61 +43,12 @@ export default function SingleChoice({
         imageLoading="eager"
       />
 
-      <div
-        className="flex flex-col gap-3"
-        role="radiogroup"
-        aria-label={groupAriaLabel}
-      >
-        {normalizedAnswers.map((answer, i) => {
-          const imageUrl = answer.image?.url?.trim();
-          const imageThumbUrl = answer.image?.thumbUrl?.trim();
-          const hasImage = Boolean(imageUrl || imageThumbUrl);
-          const label = answer.text?.trim() || `Odpowiedź ${i + 1}`;
-
-          if (!hasImage) {
-            return (
-              <RadioAnswer
-                key={`${i}-${label}`}
-                label={label}
-                selected={selected === i}
-                onClick={() => setSelected(i)}
-              />
-            );
-          }
-
-          return (
-            <button
-              key={`${i}-${label}`}
-              type="button"
-              role="radio"
-              aria-checked={selected === i}
-              aria-label={label}
-              onClick={() => setSelected(i)}
-              className={cn(
-                'cursor-pointer rounded-2xl border p-3 text-left transition-colors',
-                selected === i
-                  ? 'border-[var(--selected-border)] bg-[var(--selected-bg)]'
-                  : 'border-[var(--border)] bg-[var(--card-bg)]',
-              )}
-            >
-              <ProgressiveQuizImage
-                thumbUrl={imageThumbUrl}
-                fullUrl={imageUrl || imageThumbUrl || ''}
-                width={answer.image?.width}
-                height={answer.image?.height}
-                alt={answer.image?.alt || label}
-                loading="lazy"
-                className="w-full rounded-xl bg-white object-contain"
-              />
-              {answer.text?.trim() && (
-                <p className="mt-2 text-sm font-medium text-[var(--text-dark)]">
-                  {answer.text.trim()}
-                </p>
-              )}
-            </button>
-          );
-        })}
-      </div>
+      <SingleSelectAnswers
+        answers={answers}
+        selected={selected}
+        onSelect={setSelected}
+        groupAriaLabel={groupAriaLabel}
+      />
 
       <SubmitButton
         label="Zatwierdź odpowiedź"
